FIX: arreglada relación subasta usuario

+++ src/main/java/es/ucm/fdi/iw/business/model/Puja.java
@@ -33,7 +33,7 @@ public class Puja {
     @JoinColumn(name = "subasta_id", nullable = false)
     private Subasta subasta;
 
-    private BigDecimal dineroPujado; 
+    private double dineroPujado; 
     private int puntuacion;
     private String comentario;
     private LocalDateTime fecha;
